setup use recruit data hook
